% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tds.R
\name{browse}
\alias{browse}
\title{Browse a catalog page if interactive}
\usage{
browse(x = cefi_top(), as = c("xml", "html"))
}
\arguments{
\item{x}{a CatalogNode object}

\item{as}{chr one of "xml" or "html"}
}
\value{
NULL invisibly
}
\description{
Browse a catalog page if interactive
}
