% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tds.R
\name{cefi_region}
\alias{cefi_region}
\title{Get a TopCatalog for a specific region}
\usage{
cefi_region(region = "northwest_atlantic", domain = "full_domain", ...)
}
\arguments{
\item{region}{chr one of "arctic", "great_lakes", "northeast_pacific"
"northwest_atlantic" or "pacific_islands"}

\item{domain}{chr, the name of the dataset catalog to retrieve
from the regional catalog set to NULL or NA to skip}

\item{...}{other arguments for \code{cefi_top()}}
}
\value{
a TopCatalog object
}
\description{
Get a TopCatalog for a specific region
}
