% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CEFI_REGIONS}
\alias{CEFI_REGIONS}
\title{A look-up-table of regional names and abbreviations}
\format{
a named character vector \code{abb = long_name}
}
\source{
\url{https://psl.noaa.gov/cefi_portal/}
}
\usage{
CEFI_REGIONS
}
\description{
A look-up-table of regional names and abbreviations
}
\keyword{datasets}
