% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tds.R
\docType{data}
\name{CEFI_SUBDOMAIN}
\alias{CEFI_SUBDOMAIN}
\title{A look-up-table of domain names and abbreviations}
\format{
a named character vector \code{abb = long_name}
}
\source{
\url{https://psl.noaa.gov/cefi_portal/}
}
\usage{
CEFI_SUBDOMAIN
}
\description{
A look-up-table of domain names and abbreviations
}
\keyword{datasets}
