% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tds.R
\name{cefi_top}
\alias{cefi_top}
\title{Get the cefi top level catalog}
\usage{
cefi_top(uri = top_uri())
}
\arguments{
\item{uri}{chr, the URI of the top level catalog}
}
\value{
TopCatalog object
}
\description{
Get the cefi top level catalog
}
