% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cefitds-package.R
\docType{package}
\name{cefitds}
\alias{cefitds-package}
\alias{cefitds}
\title{cefitds package}
\description{
R tools access CEFI's THREDDS catalogs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bigelowlab/cefitds}
}

}
