% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cefi.R
\name{match_catalog_table}
\alias{match_catalog_table}
\title{Match a cefi_dataset_table to a CEFI_catalog table}
\usage{
match_catalog_table(x = parse_url())
}
\arguments{
\item{x}{a cefi_dataset_table object}
}
\value{
a CEFI_catalog matching the rows of \code{x}
}
\description{
Match a cefi_dataset_table to a CEFI_catalog table
}
