% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tds.R
\name{node_extract_table}
\alias{node_extract_table}
\title{Given one or more DatasetNode objects, extract a table}
\usage{
node_extract_table(
  x = query_cefi(),
  stub = "http://psl.noaa.gov/thredds/dodsC"
)
}
\arguments{
\item{x}{a list of one or more DatasetNodes}

\item{stub}{chr, the base URL stub ofr opendap service}
}
\value{
a table of class "cefi_dataset_table"
}
\description{
Given one or more DatasetNode objects, extract a table
}
