% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tds.R
\name{node_extract_url}
\alias{node_extract_url}
\title{Given one or more DatasetNode objects, extract the URLs}
\usage{
node_extract_url(x = query_cefi(), stub = "http://psl.noaa.gov/thredds/dodsC")
}
\arguments{
\item{x}{a list of one or more DatasetNodes}

\item{stub}{chr, the base URL stub ofr opendap service}
}
\value{
character vector of URLs for NetCDF resources
}
\description{
Given one or more DatasetNode objects, extract the URLs
}
