% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cefi.R
\name{parse_url}
\alias{parse_url}
\title{Parse a cefi URL into a table}
\usage{
parse_url(
  x = c("sfc_co3_sol_calc.nwa.full.hcast.monthly.raw.r20230520.199301-201912.nc",
    "eparag100.nwa.full.hcast.monthly.raw.r20230520.199301-201912.nc",
    "sfc_no3lim_lgp.nwa.full.hcast.monthly.raw.r20230520.199301-201912.nc",
    "sfc_nh4lim_lgp.nwa.full.hcast.monthly.regrid.r20230520.199301-201912.nc",
    "hfevapds.nwa.full.hcast.monthly.raw.r20230520.199301-201912.nc")
)
}
\arguments{
\item{x}{one or more URLs}
}
\value{
table
}
\description{
Parse a cefi URL into a table
}
