% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tds.R
\name{select_datasets}
\alias{select_datasets}
\title{Filter the datasets for ones starting with the specified
variable names}
\usage{
select_datasets(x = query_cefi(), vars = c("btm_o2", "btm_temp"))
}
\arguments{
\item{x}{TopCatalog with datasets}

\item{vars}{chr one or more \code{short_name} variable names}
}
\value{
list of one or more DatasetNode object
}
\description{
Filter the datasets for ones starting with the specified
variable names
}
