% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tds.R
\name{top_uri}
\alias{top_uri}
\title{Retrieve the top level URI}
\usage{
top_uri(ext = c(".xml", ".html")[1])
}
\arguments{
\item{ext}{chr the extension to use (".xml" or ".html")}
}
\value{
URI of the top level catalog
}
\description{
Retrieve the top level URI
}
